requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
base58==2.1.1
bitarray==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.7
ckzg==2.1.7
click==8.3.2
coincurve==21.0.0
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
fastapi==0.135.3
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.1
parsimonious==0.10.0
pluggy==1.6.0
pycryptodome==3.23.0
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rlp==4.1.0
ruff==0.15.10
starlette==1.0.0
toolz==1.1.0
-e git+https://github.com/hshadab/tron.git@b42e9360e8dd0415edf04aa837037905c8298b23#egg=tron_preflight_demo
tronpy==0.6.2
tvm-x402==0.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
